Implemented task a) in lab1
If we don't use sleep() in while(true)-loop at thread's run() method, all seems to be working as it expected, but it loads computer very much.
But if we use sleep(), it seems threads have same priority (because slider has the same time at left and right), even if only one core enabled for this process.